requirements.txt
flask
discord.py
tinydb